Clean up Vertex import
