+++ src/shaders/colored_triangle.vert
@@ -7,11 +7,12 @@ layout (push_constant) uniform block {
 };
 layout (location = 0) in vec3 inPosition;
 layout (location = 1) in vec3 inColour;
-layout (location = 2) in vec2 inUV;
+layout (location = 2) in vec3 inNormal;
+layout (location = 3) in vec2 inUV;
 
 layout (location = 0) out vec3 outColor;
 
 void main() {
     gl_Position = projection * view * model * vec4(inPosition, 1.0);
-    outColor = inColour;
+    outColor = normalize(inNormal + vec3(0., 0., 2.));
 }
